requirements.txt
streamlit==1.42.1

# Nuevo SDK OpenAI 1.x
openai==1.63.2

# Pinecone
pinecone==5.4.1

python-dotenv==1.0.1
reportlab==4.3.1
joblib==1.4.2
tqdm==4.67.1
psutil==7.0.0
scikit-learn==1.6.1
